feat: custom hotkey for Activation Shortcut (#44)
You can now set the Activation Shortcut to any combination of SHIFT, CONTROL, OPTION and COMMAND along with SPACE. It's still not possible to customize the SPACE.

Additionally, this update removes the settings for capturing an OpenAI and cross.stream token. Users who already have a token will still be able to sync a single stack to cross.stream. Going forward, cross.stream syncing will not be directly integrated but will be accessible through an addon plugin or similar.

<img width="435" alt="image" src="https://github.com/cablehead/stacks/assets/1394/85e7257e-21fe-4a79-9521-538e7fdc7f01">